rename files and fmt
