+++ lib/block/account_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"boscoin.io/sebak/lib/common"
-	"boscoin.io/sebak/lib/observer"
+	"boscoin.io/sebak/lib/common/observer"
 	"boscoin.io/sebak/lib/storage"
 
 	"github.com/stretchr/testify/require"
