+++ lib/network/memory.go
