+++ lib/network/memory_client.go
