+++ lib/node/runner/api_node.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"io/ioutil"
