+++ lib/node/runner/api_node_test.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"fmt"
