+++ lib/node/runner/checker_message_test.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"testing"
