+++ lib/node/runner/constant.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"time"
