+++ lib/node/runner/init.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"os"
