+++ lib/node/runner/node_runner_memory_network_test.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"boscoin.io/sebak/lib/common"
