+++ lib/node/runner/node_runner_test.go
@@ -1,4 +1,4 @@
-package noderunner
+package runner
 
 import (
 	"fmt"
